requirements.txt
matplotlib
pandas
pyserial